+++ CHANGELOG.md
@@ -10,3 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a `CpuId::with_cpuid_fn` that allows to override the default cpuid function.
 - Refactored tests into a module.
 - Started to add tests for Ryzen/AMD.
+- Fixed `RdtAllocationInfo.has_memory_bandwidth_allocation`: was using the wrong bit
+- Fixed `capacity_mask_length` in `L3CatInfo` and `L2CatInfo`: add +1 to returned value
+- Fixed `MemBwAllocationInfo.max_hba_throttling`: add +1 to returned value